+++ example/basic_http_interceptor_hook_example.dart
@@ -1,5 +1,4 @@
 import 'package:logging/logging.dart';
-import 'package:http_interceptor/http_interceptor.dart';
 
 import 'package:basic_http_interceptor/basic_http_interceptor.dart';