+++ lib/basic_http_interceptor.dart
@@ -1,6 +1,5 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
+/// Lightweight, pluggable HTTP interceptors for logging, default headers,
+/// functional hooks, proxy configurations, and request timeout abortion.
 library basic_http_interceptor;
 
 import 'dart:async';
@@ -10,17 +9,20 @@ import 'dart:typed_data';
 import 'package:http_interceptor/extensions/streamed_response.dart';
 import 'package:http_interceptor/http_interceptor.dart'
     show
+        Abortable,
+        AbortableMultipartRequest,
+        AbortableRequest,
         BaseRequest,
         BaseResponse,
         InterceptorContract,
+        MultipartRequest,
         Request,
         Response,
         StreamedResponse;
 
 import 'package:logging/logging.dart';
 
-export 'package:http_interceptor/http_interceptor.dart'
-    show InterceptedClient, InterceptorContract;
+export 'package:http_interceptor/http_interceptor.dart';
 
 export 'package:http_interceptor/utils/query_parameters.dart'
     show buildUrlString;
@@ -32,6 +34,7 @@ export 'utils/basic_http_method_switching_client.dart';
 part 'src/basic_http_interceptor_logger.dart';
 part 'src/basic_http_interceptor_header.dart';
 part 'src/basic_http_interceptor_hook.dart';
+part 'src/basic_http_interceptor_timeout.dart';
 
 // export 'src/basic_http_interceptor_base.dart';